Rename and separate files for top sites
